+++ src/environments/environment.ts
@@ -9,7 +9,7 @@ export const environment = {
   production: false,
   stripePublicKey: 'pk_test_5D57doatt7FpFolXlTHy8DvZ00TgL3h1WJ',
   url: 'http://localhost:4200',
-  version: 'v2.2.0b',
+  version: 'v2.3.0b',
   voteURLs: [
     { name: '12 hours (1)', url: '//dbots.co/bots/525935335918665760/vote' },
     { name: '12 hours (2)', url: '//top.gg/bot/525935335918665760/vote' },