+++ src/assets/docs/changelog.md
@@ -12,16 +12,18 @@
 
 ---
 
-### v2.2.1b
+### v2.3.0b
+`10/09/2020`
 
 **Add**
+- `/servers/:id/commands` -> custom commands
 
 **Change/Fix**
 - 
 
 **Remove**
-- dashboard -> badges
-- dashboard -> crates (may be temporary)
+- `/dashboard` -> badges
+- `/dashboard` -> crates (may be temporary)
 
 ---